STB-4382 - Refactored to keep user activation in a separate controller, so UserController is cleaner and manageable
