+++ src/main/resources/templates/delegate-reactivate-user-reason.html
@@ -12,7 +12,7 @@
 <body>
 <nav aria-label="Breadcrumb" class="govuk-breadcrumbs" id="breadcrumbs">
     <a class="govuk-back-link govuk-!-margin-bottom-3"
-       th:href="@{/admin/users/delegate-reactivate/{id}(id=${user.id}, referer=${referer}, profileId=${profileId})}">
+       th:href="@{/admin/user/delegate-reactivate/{id}(id=${user.id}, referer=${referer}, profileId=${profileId})}">
         Back </a>
 </nav>
 <main class="govuk-main-wrapper" id="main-content">
