+++ src/main/resources/templates/manage-user.html
@@ -256,7 +256,7 @@
                                             <a th:if="${canEnableUser}" class="govuk-link"
                                                th:href="@{/admin/users/manage/{id}/enable(id=${user.entraUser.id}, referer='manage', profileId=${user.id})}">Activate user</a>
                                             <a th:if="${canDelegateEnableUser}" class="govuk-link"
-                                               th:href="@{/admin/users/delegate-reactivate/{id}(id=${user.entraUser.id}, referer='manage', profileId=${user.id})}">Reactivate user</a>
+                                               th:href="@{/admin/user/delegate-reactivate/{id}(id=${user.entraUser.id}, referer='manage', profileId=${user.id})}">Reactivate user</a>
                                             <span th:if="${cannotEnableUser}">You cannot activate this user</span>
                                         </dd>
                                     </div>
