+++ src/test/java/uk/gov/justice/laa/portal/landingpage/controller/UserControllerTest.java
@@ -19,10 +19,8 @@ import java.util.stream.Collectors;
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Assertions;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -37,7 +35,6 @@ import static org.mockito.ArgumentMatchers.eq;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 
-import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.lenient;
@@ -59,7 +56,6 @@ import org.springframework.ui.ExtendedModelMap;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.ObjectError;
-import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.mvc.support.RedirectAttributesModelMap;
 import org.springframework.web.servlet.view.RedirectView;
@@ -95,8 +91,6 @@ import uk.gov.justice.laa.portal.landingpage.entity.FirmType;
 import uk.gov.justice.laa.portal.landingpage.entity.InvitationStatus;
 import uk.gov.justice.laa.portal.landingpage.entity.Office;
 import uk.gov.justice.laa.portal.landingpage.entity.Permission;
-import uk.gov.justice.laa.portal.landingpage.entity.ReactivationRequestStatus;
-import uk.gov.justice.laa.portal.landingpage.entity.UserActivationRequest;
 import uk.gov.justice.laa.portal.landingpage.entity.UserProfile;
 import uk.gov.justice.laa.portal.landingpage.entity.UserProfileSilasStatus;
 import uk.gov.justice.laa.portal.landingpage.entity.UserProfileStatus;
@@ -106,7 +100,6 @@ import uk.gov.justice.laa.portal.landingpage.entity.UserTypeReasonDisable;
 import uk.gov.justice.laa.portal.landingpage.exception.CreateUserDetailsIncompleteException;
 import uk.gov.justice.laa.portal.landingpage.exception.TechServicesClientException;
 import uk.gov.justice.laa.portal.landingpage.forms.ApplicationsForm;
-import uk.gov.justice.laa.portal.landingpage.forms.DelegateReactivateUserReasonForm;
 import uk.gov.justice.laa.portal.landingpage.forms.DisableUserReasonForm;
 import uk.gov.justice.laa.portal.landingpage.forms.EditUserDetailsForm;
 import uk.gov.justice.laa.portal.landingpage.forms.FirmSearchForm;
@@ -129,7 +122,6 @@ import uk.gov.justice.laa.portal.landingpage.service.NotificationService;
 import uk.gov.justice.laa.portal.landingpage.service.OfficeService;
 import uk.gov.justice.laa.portal.landingpage.service.RoleAssignmentService;
 import uk.gov.justice.laa.portal.landingpage.service.UserAccountStatusService;
-import uk.gov.justice.laa.portal.landingpage.service.UserReactivationActivationRequestService;
 import uk.gov.justice.laa.portal.landingpage.service.UserService;
 import uk.gov.justice.laa.portal.landingpage.techservices.SendUserVerificationEmailResponse;
 import uk.gov.justice.laa.portal.landingpage.techservices.TechServicesApiResponse;
@@ -175,20 +167,14 @@ class UserControllerTest {
     private UserAccountStatusService disableUserService;
     @Mock
     private NotificationService notificationService;
-    @Mock
-    private UserReactivationActivationRequestService userReactivationActivationRequestService;
     private Model model;
     private FirmSearchForm firmSearchForm;
-    private static final String USER_ID = "entra-user-123";
-    private static final UUID PROFILE_ID = UUID.randomUUID();
-    private static final String REFERER = "http://localhost/admin/users";
-
 
     @BeforeEach
     void setUp() {
         userController = new UserController(loginService, userService, officeService, eventService, firmService,
                 new MapperConfig().modelMapper(), accessControlService, roleAssignmentService, emailValidationService,
-                appRoleService, appService, disableUserService, notificationService, userReactivationActivationRequestService);
+                appRoleService, appService, disableUserService, notificationService);
         userController.disableUserFeatureEnabled = true;
         lenient().when(accessControlService.getEnablementFlags(any()))
                 .thenReturn(new AccessControlService.EnablementFlags(false, false, false));
@@ -821,7 +807,7 @@ class UserControllerTest {
         List<uk.gov.justice.laa.portal.landingpage.viewmodel.DeleteUserReasonViewModel> reasons =
                 (List<uk.gov.justice.laa.portal.landingpage.viewmodel.DeleteUserReasonViewModel>) model.getAttribute("deleteReasons");
         assertThat(reasons).hasSize(1);
-        assertThat(reasons.get(0).getCode()).isEqualTo("CyberRisk");
+        assertThat(reasons.getFirst().getCode()).isEqualTo("CyberRisk");
         verify(userService).getDeleteUserReasons(true);
     }
 
@@ -8435,338 +8421,6 @@ class UserControllerTest {
                     .isEqualTo("Failed to generate and send activation code via email.");
         }
 
-    }
-
-    private EntraUserDto buildEntraUserDto() {
-        return EntraUserDto.builder()
-                .id(USER_ID)
-                .fullName("John Doe")
-                .email("john.doe@justice.gov.uk")
-                .build();
-    }
-
-    private UserActivationRequest buildUserActivationRequest(ReactivationRequestStatus status) {
-        return UserActivationRequest.builder()
-                .id(UUID.randomUUID())
-                .requestId(UUID.randomUUID())
-                .userProfileId(PROFILE_ID)
-                .status(status)
-                .version(1)
-                .build();
-    }
-
-    // ==========================================
-    // GET /users/delegate-reactivate/{id}
-    // ==========================================
-    @Nested
-    @DisplayName("GET /users/delegate-reactivate/{id}")
-    class DelegateReactivateUserGetTests {
-
-        @Test
-        @DisplayName("Should throw 404 when feature flag is disabled")
-        void featureDisabled_throws404() {
-            userController.disableUserFeatureEnabled = false;
-
-            assertThatThrownBy(() -> userController.delegateReactivateUserGet(USER_ID, session, model, REFERER, PROFILE_ID, redirectAttributes))
-                    .isInstanceOf(ResponseStatusException.class)
-                    .hasMessageContaining("404");
-        }
-
-        @Test
-        @DisplayName("Should redirect with error when request is already in progress (PENDING status)")
-        void requestInProgress_redirectsToManagePage() {
-            EntraUserDto user = buildEntraUserDto();
-            UserActivationRequest pendingRequest = buildUserActivationRequest(ReactivationRequestStatus.IN_REVIEW);
-
-            given(userService.getEntraUserById(USER_ID)).willReturn(Optional.of(user));
-            given(userReactivationActivationRequestService.findFirstByUserProfileIdOrderByVersionDesc(PROFILE_ID))
-                    .willReturn(Optional.of(pendingRequest));
-
-            redirectAttributes = new RedirectAttributesModelMap();
-            String view = userController.delegateReactivateUserGet(USER_ID, session, model, REFERER, PROFILE_ID, redirectAttributes);
-
-            assertThat(view).isEqualTo("redirect:/admin/users/manage/" + PROFILE_ID);
-            assertThat(redirectAttributes.getFlashAttributes())
-                    .extractingByKey("errorMessage")
-                    .isEqualTo("A delegate request is already in progress");
-            assertThat(redirectAttributes.getFlashAttributes())
-                    .extractingByKey("requestId")
-                    .isEqualTo(pendingRequest.getId());
-        }
-
-        @Test
-        @DisplayName("Should populate model and session when no active request exists (or is REJECTED/APPROVED)")
-        void noActiveRequest_rendersForm() {
-            session = new MockHttpSession();
-            EntraUserDto user = buildEntraUserDto();
-            UserActivationRequest rejectedRequest = buildUserActivationRequest(ReactivationRequestStatus.REJECTED);
-
-            given(userService.getEntraUserById(USER_ID)).willReturn(Optional.of(user));
-            given(userReactivationActivationRequestService.findFirstByUserProfileIdOrderByVersionDesc(PROFILE_ID))
-                    .willReturn(Optional.of(rejectedRequest));
-
-            String view = userController.delegateReactivateUserGet(USER_ID, session, model, REFERER, PROFILE_ID, redirectAttributes);
-
-            assertThat(view).isEqualTo("delegate-reactivate-user");
-            assertThat(model.asMap())
-                    .containsEntry("user", user)
-                    .containsEntry("profileId", PROFILE_ID)
-                    .containsEntry("referer", REFERER);
-
-            assertThat(session.getAttribute("delegateReactivateUserId")).isEqualTo(USER_ID);
-            assertThat(session.getAttribute("profileId")).isEqualTo(PROFILE_ID);
-        }
-    }
-
-    // ==========================================
-    // POST /users/delegate-reactivate/{id}
-    // ==========================================
-    @Nested
-    @DisplayName("POST /users/delegate-reactivate/{id}")
-    class DelegateReactivateUserPostTests {
-
-        @Test
-        @DisplayName("Should throw 403 when session ID does not match path variable")
-        void mismatchSessionId_throws403() {
-            session = new MockHttpSession();
-            session.setAttribute("delegateReactivateUserId", "different-user-id");
-
-            assertThatThrownBy(() -> userController.delegateReactivateUserPost(USER_ID, model, session, REFERER, PROFILE_ID))
-                    .isInstanceOf(ResponseStatusException.class)
-                    .hasMessageContaining("403");
-        }
-
-        @Test
-        @DisplayName("Should set attributes and redirect to reasons step on success")
-        void validSession_redirectsToReasons() {
-            session = new MockHttpSession();
-            session.setAttribute("delegateReactivateUserId", USER_ID);
-            EntraUserDto user = buildEntraUserDto();
-
-            given(userService.getEntraUserById(USER_ID)).willReturn(Optional.of(user));
-
-            String view = userController.delegateReactivateUserPost(USER_ID, model, session, REFERER, PROFILE_ID);
-
-            assertThat(view).isEqualTo("redirect:/admin/users/delegate-reactivate-user-reason/" + USER_ID);
-            assertThat(session.getAttribute("delegateReactivateUserId")).isEqualTo(USER_ID);
-            assertThat(session.getAttribute("profileId")).isEqualTo(PROFILE_ID);
-        }
-    }
-
-    // ==========================================
-    // GET /users/delegate-reactivate-user-reason/{id}
-    // ==========================================
-    @Nested
-    @DisplayName("GET /users/delegate-reactivate-user-reason/{id}")
-    class DelegateReactivateUserReasonsGetTests {
-
-        @Test
-        @DisplayName("Should throw 404 when feature flag is disabled")
-        void featureDisabled_throws404() {
-            userController.disableUserFeatureEnabled = false;
-
-            assertThatThrownBy(() -> userController.delegateReactivateUserReasonsGet(USER_ID, model, session))
-                    .isInstanceOf(ResponseStatusException.class)
-                    .hasMessageContaining("404");
-        }
-
-        @Test
-        @DisplayName("Should throw 403 when session user ID does not match path variable")
-        void invalidSessionUser_throws403() {
-            session = new MockHttpSession();
-            session.setAttribute("delegateReactivateUserId", "invalid-id");
-
-            assertThatThrownBy(() -> userController.delegateReactivateUserReasonsGet(USER_ID, model, session))
-                    .isInstanceOf(ResponseStatusException.class)
-                    .hasMessageContaining("403");
-        }
-
-        @Test
-        @DisplayName("Should load form from session if present or create new instance")
-        void validSession_rendersReasonsView() {
-            session = new MockHttpSession();
-            session.setAttribute("delegateReactivateUserId", USER_ID);
-            session.setAttribute("profileId", PROFILE_ID);
-
-            DelegateReactivateUserReasonForm existingForm = DelegateReactivateUserReasonForm.builder()
-                    .reason("Existing reason")
-                    .build();
-            session.setAttribute("delegateReactivateUserReasonForm", existingForm);
-
-            EntraUserDto user = buildEntraUserDto();
-            given(userService.getEntraUserById(USER_ID)).willReturn(Optional.of(user));
-
-            String view = userController.delegateReactivateUserReasonsGet(USER_ID, model, session);
-
-            assertThat(view).isEqualTo("delegate-reactivate-user-reason");
-            assertThat(model.asMap())
-                    .containsEntry("user", user)
-                    .containsEntry("profileId", PROFILE_ID)
-                    .containsEntry("delegateReactivateUserReasonForm", existingForm);
-        }
-    }
-
-    // ==========================================
-    // POST /users/delegate-reactivate-user-reason/{id}
-    // ==========================================
-    @Nested
-    @DisplayName("POST /users/delegate-reactivate-user-reason/{id}")
-    class DelegateReactivateUserReasonsPostTests {
-
-        @Test
-        @DisplayName("Should re-render form with error message when validation fails")
-        void validationErrors_returnsReasonViewWithError() {
-            session = new MockHttpSession();
-            session.setAttribute("profileId", PROFILE_ID);
-            DelegateReactivateUserReasonForm form = DelegateReactivateUserReasonForm.builder().build();
-
-            BindingResult bindingResult = Mockito.mock(BindingResult.class);
-            given(bindingResult.hasErrors()).willReturn(true);
-
-            EntraUserDto user = buildEntraUserDto();
-            given(userService.getEntraUserById(USER_ID)).willReturn(Optional.of(user));
-
-            String view = userController.delegateReactivateUserReasonsPost(USER_ID, form, bindingResult, model, session);
-
-            assertThat(view).isEqualTo("delegate-reactivate-user-reason");
-            assertThat(model.asMap()).containsKey("errorMessage");
-            assertThat(session.getAttribute("delegateReactivateUserReasonForm")).isNull();
-        }
-
-        @Test
-        @DisplayName("Should save form to session and redirect to check answers when validation passes")
-        void validationSuccess_savesSessionAndRedirects() {
-            session = new MockHttpSession();
-            session.setAttribute("profileId", PROFILE_ID);
-            DelegateReactivateUserReasonForm form = DelegateReactivateUserReasonForm.builder()
-                    .reason("Valid user reactivation request reason")
-                    .build();
-
-            BindingResult bindingResult = Mockito.mock(BindingResult.class);
-            given(bindingResult.hasErrors()).willReturn(false);
-            EntraUserDto user = buildEntraUserDto();
-            given(userService.getEntraUserById(USER_ID)).willReturn(Optional.of(user));
-
-            String view = userController.delegateReactivateUserReasonsPost(USER_ID, form, bindingResult, model, session);
-
-            assertThat(view).isEqualTo("redirect:/admin/users/delegate-reactivate-user-check-answers/" + USER_ID);
-            assertThat(session.getAttribute("delegateReactivateUserReasonForm")).isEqualTo(form);
-        }
-    }
-
-    // ==========================================
-    // Builder Helpers
-    // ==========================================
-
-    // ==========================================
-    // GET /users/delegate-reactivate-user-check-answers/{id}
-    // ==========================================
-    @Nested
-    @DisplayName("GET /users/delegate-reactivate-user-check-answers/{id}")
-    class DelegateReactivateUserReasonsCheckAnswersGetTests {
-
-        @Test
-        @DisplayName("Should throw 403 if session ID does not match path variable")
-        void invalidSessionUser_throws403() {
-            session = new MockHttpSession();
-            session.setAttribute("profileId", PROFILE_ID);
-            session.setAttribute("delegateReactivateUserId", "mismatched-id");
-
-            assertThatThrownBy(() -> userController.delegateReactivateUserReasonsCheckAnswersGet(USER_ID, model, session))
-                    .isInstanceOf(ResponseStatusException.class)
-                    .hasMessageContaining("403");
-        }
-
-        @Test
-        @DisplayName("Should throw NoSuchElementException if form is missing from session")
-        void missingFormInSession_throwsException() {
-            session.setAttribute("profileId", PROFILE_ID);
-            session.setAttribute("delegateReactivateUserId", USER_ID);
-
-            assertThatThrownBy(() -> userController.delegateReactivateUserReasonsCheckAnswersGet(USER_ID, model, session))
-                    .isInstanceOf(NoSuchElementException.class);
-        }
-
-        @Test
-        @DisplayName("Should render check answers view with populated model")
-        void validSession_rendersCheckAnswersView() {
-            session = new MockHttpSession();
-            session.setAttribute("profileId", PROFILE_ID);
-            session.setAttribute("delegateReactivateUserId", USER_ID);
-
-            DelegateReactivateUserReasonForm form = DelegateReactivateUserReasonForm.builder().reason("Reactivation justification").build();
-            session.setAttribute("delegateReactivateUserReasonForm", form);
-
-            EntraUserDto user = buildEntraUserDto();
-            given(userService.getEntraUserById(USER_ID)).willReturn(Optional.of(user));
-
-            String view = userController.delegateReactivateUserReasonsCheckAnswersGet(USER_ID, model, session);
-
-            assertThat(view).isEqualTo("delegate-reactivate-user-check-answers");
-            assertThat(model.asMap())
-                    .containsEntry("user", user)
-                    .containsEntry("profileId", PROFILE_ID)
-                    .containsEntry("delegateReactivateUserReasonForm", form);
-        }
-    }
-
-    // ==========================================
-    // POST /users/delegate-reactivate-user-check-answers/{id}
-    // ==========================================
-    @Nested
-    @DisplayName("POST /users/delegate-reactivate-user-check-answers/{id}")
-    class DelegateReactivateUserReasonsCheckAnswersPostTests {
-
-        @Test
-        @DisplayName("Should throw 400 if user profile ID does not match session profile ID")
-        void profileIdMismatch_throws400() {
-            session = new MockHttpSession();
-            session.setAttribute("delegateReactivateUserId", USER_ID);
-            session.setAttribute("profileId", PROFILE_ID);
-            session.setAttribute("delegateReactivateUserReasonForm", DelegateReactivateUserReasonForm.builder().build());
-
-            UserProfileDto mismatchedProfile = UserProfileDto.builder().id(UUID.randomUUID()).build();
-            given(userService.getActiveProfileByUserId(USER_ID)).willReturn(Optional.of(mismatchedProfile));
-
-            assertThatThrownBy(() -> userController.delegateReactivateUserReasonsCheckAnswersPost(USER_ID, authentication, model, session))
-                    .isInstanceOf(ResponseStatusException.class)
-                    .hasMessageContaining("400");
-
-            verify(userReactivationActivationRequestService, never()).createNewRequest(any(), any(), any(), any());
-        }
-
-        @Test
-        @DisplayName("Should create activation request and render confirmation view on success")
-        void validSubmission_createsRequestAndRendersConfirmation() {
-            MockHttpSession httpSession = new MockHttpSession();
-            httpSession.setAttribute("delegateReactivateUserId", USER_ID);
-            httpSession.setAttribute("profileId", PROFILE_ID);
-
-            DelegateReactivateUserReasonForm form = DelegateReactivateUserReasonForm.builder()
-                    .reason("Approved leave returned")
-                    .build();
-            httpSession.setAttribute("delegateReactivateUserReasonForm", form);
-
-            UserProfileDto userProfile = UserProfileDto.builder().id(PROFILE_ID).build();
-            EntraUserDto user = buildEntraUserDto();
-            CurrentUserDto actor = new CurrentUserDto();
-            actor.setUserId(UUID.randomUUID());
-            UserActivationRequest createdRequest = buildUserActivationRequest(ReactivationRequestStatus.IN_REVIEW);
-
-            given(userService.getActiveProfileByUserId(USER_ID)).willReturn(Optional.of(userProfile));
-            given(userService.getEntraUserById(USER_ID)).willReturn(Optional.of(user));
-            given(loginService.getCurrentUser(authentication)).willReturn(actor);
-            given(userReactivationActivationRequestService.createNewRequest(any(UUID.class), eq(PROFILE_ID), eq("Approved leave returned"), eq(user)))
-                    .willReturn(createdRequest);
-
-            String view = userController.delegateReactivateUserReasonsCheckAnswersPost(USER_ID, authentication, model, httpSession);
-
-            assertThat(view).isEqualTo("delegate-reactivate-user-confirmation");
-            assertThat(model.asMap()).containsEntry("user", user);
-
-            verify(userReactivationActivationRequestService)
-                    .createNewRequest(any(UUID.class), eq(PROFILE_ID), eq("Approved leave returned"), eq(user));
-        }
     }
 
 }
